refactor: domain 패키지 내 Wish, Member, Product 클래스를 @Entity로 설정
