+++ src/main/java/gift/domain/Wish.java
@@ -1,23 +1,27 @@
 package gift.domain;
 
-import java.time.LocalDateTime;
+import jakarta.persistence.*;
 
-public class Wish {
+@Entity
+@Table(name = "wish")
+public class Wish extends BaseEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private Long memberId;
-    private Long productId;
-    private LocalDateTime createdAt;
+
+    @ManyToOne
+    @JoinColumn(name = "member_id", nullable = false)
+    private Member member;
+
+    @ManyToOne
+    @JoinColumn(name = "product_id", nullable = false)
+    private Product product;
 
     public Long getMemberId() {
-        return memberId;
+        return member.getId();
     }
 
     public Long getProductId() {
-        return productId;
-    }
-
-    public Wish(Long memberId, Long productId) {
-        this.memberId = memberId;
-        this.productId = productId;
+        return product.getId();
     }
 }
