--- src/main/java/gift/entity/Member.java
@@ -1,21 +1,0 @@
-package gift.entity;
-
-import gift.enums.Role;
-import jakarta.persistence.*;
-
-@Entity
-@Table(name = "member")
-public class Member extends BaseEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false, unique = true)
-    private String email;
-
-    @Column(nullable = false)
-    private String password;
-
-    @Column(nullable = false)
-    private Role role;
-}
