--- src/main/java/gift/entity/Product.java
@@ -1,20 +1,0 @@
-package gift.entity;
-
-import jakarta.persistence.*;
-
-@Entity
-@Table(name = "product")
-public class Product extends BaseEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String name;
-
-    @Column(nullable = false)
-    private Long price;
-
-    @Column(nullable = false)
-    private String imageUrl;
-}
