--- src/main/java/gift/entity/Wish.java
@@ -1,19 +1,0 @@
-package gift.entity;
-
-import jakarta.persistence.*;
-
-@Entity
-@Table(name = "wish")
-public class Wish extends BaseEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne
-    @JoinColumn(name = "member_id", nullable = false)
-    private Member member;
-
-    @ManyToOne
-    @JoinColumn(name = "product_id", nullable = false)
-    private Product product;
-}
